api/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
boto3==1.34.11
botocore==1.34.11
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.105.0
h11==0.14.0
httpx==0.26.0
idna==3.6
jmespath==1.0.1
pillow==11.0.0
pydantic==2.9.2
pypng==0.20220715.0
python-dateutil==2.8.2
python-dotenv==1.0.0
qrcode==7.4.2
s3transfer==0.10.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.12.2
urllib3==1.26.18
uvicorn==0.24.0.post1
